scan_for_focus_path should return an empty array if path is a directory
When I have set `focus_on`, and I run all tests, then a directory path may be sent to Focuser#scan_for_focus_path. This method tries to call File#gets on the path, and ruby throws a "Is a directory" error.

This patch simply returns an empty array if the file is a directory.